Moved setting up DCC_PIN to SetupDCC()
Better encapsulation of DCC_PIN's set up by moving to SetupDCC. Also, changed DCC_OutPin type declaration from int to uint8_t in SetupDCC().

+++ libraries/CommandTest/DCCLibrary.h
@@ -1,7 +1,7 @@
 #include "Arduino.h"
 
 
-void SetupDCC(int DCC_OutPin);
+void SetupDCC(uint8_t DCC_OutPin);
 
 
 // buffer for command
